+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Add support to `nostrdb` storage backend ([Yuki Kishimoto])
 * nostr: add `Report::Other` variant ([Daniel Cadenas])
+* nostr: add `EventBuilder::reaction_extended` ([Yuki Kishimoto])
 * pool: add `Relay::handle_notifications` ([Yuki Kishimoto])
 * cli: add command to serve `Nostr Connect` signer ([Yuki Kishimoto])
 * ffi(nostr): added `FilterRecord`, to allow to access fields in `Filter` ([Yuki Kishimoto])